recipes-multimedia: Use git prefix for the common submodule URL
Use git prefix for the gstreamer/common URL as
without this the git clone method was not being used
for cloning the common submodule.

Also remove SRC_URI md5sum as it is not needed for
common anymore after using the git prefix as yocto now
correctly uses the git instead of assuming the
https URL as tarball.

Signed-off-by: Devarsh Thakkar <devarsht@xilinx.com>